backend/requirements.txt
alembic==1.13.2
autopep8==2.3.1
beautifulsoup4
docling
psycopg2==2.9.9
pytest-playwright
PyMuPDF==1.25.2
pymupdf4llm==0.0.17
python-dotenv==1.0.1
sentry-sdk==2.14.0
SQLAlchemy==2.0.33
requests==2.32.2